Fixed crud function
