+++ src/main/java/com/example/AuthAndAut/controller/AuthController.java
@@ -41,51 +41,49 @@ public class AuthController {
     }
 
 
-    @PreAuthorize("isAuthenticated()")
-    @PostMapping("/addPost")
-    public ResponseEntity<?> addPost(@RequestBody Post post) {
-        if (post == null) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid or missing post data");
-        }
-
-        // Get the current authenticated user
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String currentUserUsername = authentication.getName();
-
-        // Load the user from the database
-        User currentUser = (User) userService.loadUserByUsername(currentUserUsername);
-
-        // Check if the user is not null
-        if (currentUser == null) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to get current user");
-        }
-
-        // Set the user to the post
-        post.setUser(currentUser);
-
-        // Save the post
-        Post savedPost = postService.savePost(post);
-
-        if (savedPost != null) {
-            return ResponseEntity.ok(savedPost);
-        } else {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to save post");
-        }
-    }
-
-
-
-
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/getPosts")
-    public ResponseEntity<List<Post>> getPosts() {
-        List<Post> posts = postService.getAllPosts();
-        if (posts.isEmpty()) {
-            return ResponseEntity.noContent().build();
-        }
-
-        return ResponseEntity.ok(posts);
-    }
+//    @PreAuthorize("isAuthenticated()")
+//    @PostMapping("/addPost")
+//    public ResponseEntity<?> addPost(@RequestBody Post post) {
+//        if (post == null) {
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid or missing post data");
+//        }
+//
+//        // Get the current authenticated user
+//        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//        String currentUserUsername = authentication.getName();
+//
+//        // Load the user from the database
+//        User currentUser = (User) userService.loadUserByUsername(currentUserUsername);
+//
+//        // Check if the user is not null
+//        if (currentUser == null) {
+//            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to get current user");
+//        }
+//
+//        // Set the user to the post
+//        post.setUser(currentUser);
+//
+//        // Save the post
+//        Post savedPost = postService.savePost(post);
+//
+//        if (savedPost != null) {
+//            return ResponseEntity.ok(savedPost);
+//        } else {
+//            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to save post");
+//        }
+//    }
+//
+//
+//    @PreAuthorize("isAuthenticated()")
+//    @GetMapping("/getPosts")
+//    public ResponseEntity<List<Post>> getPosts() {
+//        List<Post> posts = postService.getAllPosts();
+//        if (posts.isEmpty()) {
+//            return ResponseEntity.noContent().build();
+//        }
+//
+//        return ResponseEntity.ok(posts);
+//    }
