+++ src/main/java/com/example/AuthAndAut/repository/PostRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post,Integer>{
-    Optional<Post>findUserPostByTitle(String title);
+
 }
